+++ backend.py
@@ -1,26 +1,34 @@
-import pyrebase, flask
+from flask import Flask, render_template, request, redirect, session 
+import pyrebase
+
+app = Flask(__name__, template_folder='HNCC-ACCOUNTS')
 
 config = {
-  'apiKey': "AIzaSyA4k6MqyRnMZcutm1Ea9S5dBnF_KVFmoY8",
-  'authDomain': "hncc-accounts-545ab.firebaseapp.com",
-  'projectId': "hncc-accounts-545ab",
-  'storageBucket': "hncc-accounts-545ab.appspot.com",
-  'messagingSenderId': "273939710419",
-  'appId': "1:273939710419:web:6944c53a0f8ea7d28aef26",
-  'measurementId': "G-TP0SXVJJ01",
-  'databaseURL': ''
+  'apiKey': "AIzaSyB97jiswJvvPHH8rWytwL4mneRoVuTO9Z8",
+  'authDomain': "authentication-9973a.firebaseapp.com",
+  'projectId': "authentication-9973a",
+  'storageBucket': "authentication-9973a.appspot.com",
+  'messagingSenderId': "620596374683",
+  'appId': "1:620596374683:web:f09e04c4f34e0223f42a3e",
+  'measurementId': "G-RJLEP25CTQ",
+  'databaseURL': ""
 }
 
 firebase = pyrebase.initialize_app(config)
 auth = firebase.auth()
-db=firebase.database()
-storage=firebase.storage()
 
-#app = flask(__name__)
+app.secret_key = "Secret"
 
-email = 'teshtbhi@gmail.com'
-password = '123456'
+@app.route('/', methods = ['POST', 'GET'])
 
-user = auth.create_user_with_email_and_password(email,password)
+def index():
+    return render_template('index.html')
 
-print(user)
+@app.route('/logout')
+
+def logout():
+    pass
+
+if __name__ == '__main__':
+    app.run(debug= True)
+
